srv6: split cli.c into 2 files
Split cli.c into two files. Each file configure a specific
object (i.e. localsid and srv6 route).
