+++ modules/srv6/cli/meson.build
@@ -2,5 +2,6 @@
 # Copyright (c) 2025 Maxime Leroy, Free Mobile
 
 cli_src += files(
-  'cli.c',
+  'localsid.c',
+  'route.c',
 )
